requirements.txt
mcp[cli]
requests
google-adk
streamlit
google-generativeai
anyio
streamlit